% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-rhasdata.Rmd.
\docType{data}
\name{ages}
\alias{ages}
\title{An ages data set}
\format{
A data frame with ages of people.
\describe{
\item{id}{id of person}
\item{age}{age of person}
}
}
\source{
This data came from \code{my-raw-data.csv}
}
\usage{
ages
}
\description{
An ages data set
}
\keyword{datasets}