% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-rhasdata.Rmd.
\docType{data}
\name{mydata}
\alias{mydata}
\title{A regression data set}
\format{
A data frame with 100 rows and 2 variables:
\describe{
\item{x}{a predictor variable}
\item{y}{a response variable}
}
\source{
This data was simulated.  But usually we might cite a source here.
}
\usage{
mydata
}
\description{
A regression data set
}
\keyword{datasets}